# [compiler] optimize `op_append` with concat rhs