Instead of creating a tmp string for string_build on the rhs of append, append directly to the string:

```
$s .= $p1 . $p2 ...; // creates a tmp string for $p1 . $p2 ...
=>
$s->reserve(strlen($s) + strlen($p1) + strlen($p2) ...);
$s .= $p1;
$s .= $p2;
```

These patterns will be optimized:

	$s .= $p1 . $p2
	$s .= $p1 . $p2 . $p3
	...

Due to the fact that we compile `"<$x>"` as string build of `"<"`, `$x` and `">"`, these patterns will be optimized too:

	$s .= "<$x>";
	$s .= /* any string interpolation expr */
	...
	See Interpolated benchmark results

Since we rewrite some sprintf into a string build, these calls might benefit from this optimization too if it's append op with sprintf on the rhs.

Benchmark results:

    name                               old time/op    new time/op    delta
    Concat::Append2                       137ns ± 1%      98ns ± 1%  -28.09%  (p=0.000 n=14+15)
    Concat::Append4                       192ns ± 0%     162ns ± 3%  -15.38%  (p=0.000 n=11+15)
    Concat::Append8                       334ns ± 0%     276ns ± 0%  -17.46%  (p=0.000 n=15+14)
    Concat::Append4empty                  192ns ± 1%     141ns ± 1%  -26.38%  (p=0.000 n=13+15)
    Concat::Append4accum                  711ns ± 2%     552ns ± 1%  -22.39%  (p=0.000 n=15+14)
    Concat::Append4complex                247ns ± 1%     207ns ± 0%  -16.44%  (p=0.000 n=15+13)
    Concat::Append4optional               193ns ± 1%     161ns ± 0%  -16.64%  (p=0.000 n=14+15)
    Concat::Append4InterpolatedString     296ns ± 2%     186ns ± 0%  -37.08%  (p=0.000 n=15+9)
    Concat::Append4InterpolatedInt        249ns ± 1%     184ns ± 2%  -25.91%  (p=0.000 n=15+14)
    [Geo mean]                            253ns          194ns       -23.18%

    name                               old alloc/op   new alloc/op   delta
    Concat::Append2                       96.0B ± 0%     48.0B ± 0%  -50.00%  (p=0.000 n=15+15)
    Concat::Append4                        208B ± 0%      104B ± 0%  -50.00%  (p=0.000 n=15+15)
    Concat::Append8                        696B ± 0%      352B ± 0%  -49.43%  (p=0.000 n=15+15)
    Concat::Append4empty                   208B ± 0%      104B ± 0%  -50.00%  (p=0.000 n=15+15)
    Concat::Append4accum                 1.08kB ± 0%    0.66kB ± 0%  -38.52%  (p=0.000 n=15+15)
    Concat::Append4complex                 272B ± 0%      168B ± 0%  -38.24%  (p=0.000 n=15+15)
    Concat::Append4optional                208B ± 0%      104B ± 0%  -50.00%  (p=0.000 n=15+15)
    Concat::Append4InterpolatedString      232B ± 0%      136B ± 0%  -41.38%  (p=0.000 n=15+15)
    Concat::Append4InterpolatedInt         280B ± 0%      280B ± 0%        ~  (all equal)
    [Geo mean]                             283B           163B       -42.35%

    name                               old allocs/op  new allocs/op  delta
    Concat::Append2                        2.00 ± 0%      1.00 ± 0%        ~  (p=0.000 n=15+15)
    Concat::Append4                        2.00 ± 0%      1.00 ± 0%        ~  (p=0.000 n=15+15)
    Concat::Append8                        2.00 ± 0%      1.00 ± 0%        ~  (p=0.000 n=15+15)
    Concat::Append4empty                   2.00 ± 0%      1.00 ± 0%        ~  (p=0.000 n=15+15)
    Concat::Append4accum                   7.00 ± 0%      3.00 ± 0%  -57.14%  (p=0.000 n=15+15)
    Concat::Append4complex                 3.00 ± 0%      2.00 ± 0%        ~  (p=0.000 n=15+15)
    Concat::Append4optional                2.00 ± 0%      1.00 ± 0%        ~  (p=0.000 n=15+15)
    Concat::Append4InterpolatedString      4.00 ± 0%      2.00 ± 0%  -50.00%  (p=0.000 n=15+15)
    Concat::Append4InterpolatedInt         4.00 ± 0%      2.00 ± 0%  -50.00%  (p=0.000 n=15+15)
    [Geo mean]                             2.81           1.42       -49.25%